moving filter func to rt_diversity_filter.go
