+++ query_test.go
@@ -7,10 +7,6 @@ import (
 	"time"
 
 	tu "github.com/libp2p/go-libp2p-testing/etc"
-	"github.com/libp2p/go-libp2p/core/peer"
-	bhost "github.com/libp2p/go-libp2p/p2p/host/basic"
-	swarmt "github.com/libp2p/go-libp2p/p2p/net/swarm/testing"
-	ma "github.com/multiformats/go-multiaddr"
 
 	"github.com/stretchr/testify/require"
 )
@@ -119,144 +115,4 @@ func checkRoutingTable(a, b *IpfsDHT) bool {
 	// loop until connection notification has been received.
 	// under high load, this may not happen as immediately as we would like.
 	return a.routingTable.Find(b.self) != "" && b.routingTable.Find(a.self) != ""
-}
-
-func TestFilterPeersByIPDiversity(t *testing.T) {
-	maxIPsPerGroup := 2
-
-	ctx := context.Background()
-	host, err := bhost.NewHost(swarmt.GenSwarm(t, swarmt.OptDisableReuseport), new(bhost.HostOpts))
-	require.NoError(t, err)
-	opts := []Option{
-		RoutingTablePeerDiversityFilter(NewRTPeerDiversityFilter(host, maxIPsPerGroup, maxIPsPerGroup)),
-	}
-	dht, err := New(ctx, host, opts...)
-	require.NoError(t, err)
-	q := query{
-		dht:            dht,
-		maxIPsPerGroup: maxIPsPerGroup,
-	}
-
-	type addr struct {
-		ipv6 bool
-		addr string
-	}
-	createPeer := func(ips ...addr) *peer.AddrInfo {
-		p := &peer.AddrInfo{ID: "", Addrs: make([]ma.Multiaddr, 0, len(ips))}
-		for _, ip := range ips {
-			var a ma.Multiaddr
-			var err error
-			if ip.ipv6 {
-				a, err = ma.NewMultiaddr("/ip6/" + ip.addr + "/tcp/4001")
-			} else {
-				a, err = ma.NewMultiaddr("/ip4/" + ip.addr + "/tcp/4001")
-			}
-			require.NoError(t, err)
-			p.Addrs = append(p.Addrs, a)
-		}
-		return p
-	}
-
-	t.Run("Different IPv4 blocks", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
-			createPeer(addr{ipv6: false, addr: "2.2.2.2"}),
-			createPeer(addr{ipv6: false, addr: "3.3.3.3"}),
-			createPeer(addr{ipv6: false, addr: "4.4.4.4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Equal(t, peers, filtered)
-	})
-
-	t.Run("Same IPv4 block, but acceptable diversity", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
-			createPeer(addr{ipv6: false, addr: "1.1.2.2"}),
-			createPeer(addr{ipv6: false, addr: "3.3.3.3"}),
-			createPeer(addr{ipv6: false, addr: "4.4.4.4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Equal(t, peers, filtered)
-	})
-
-	t.Run("Overrepresented IPv4 block", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
-			createPeer(addr{ipv6: false, addr: "1.1.1.2"}),
-			createPeer(addr{ipv6: false, addr: "1.1.3.3"}),
-			createPeer(addr{ipv6: false, addr: "4.4.4.4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Len(t, filtered, 1)
-		require.Equal(t, peers[3], filtered[0])
-	})
-
-	t.Run("Different IPv6 ASNs", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: true, addr: "2001:4860:4860::1"}),
-			createPeer(addr{ipv6: true, addr: "2606:4700:4700::2"}),
-			createPeer(addr{ipv6: true, addr: "2620:fe::3"}),
-			createPeer(addr{ipv6: true, addr: "2a02:6b8::4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Equal(t, filtered, peers)
-	})
-
-	t.Run("Same IPv6 ASNs, but acceptable diversity", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: true, addr: "2001:4860:4860::1"}),
-			createPeer(addr{ipv6: true, addr: "2001:4860:4860::2"}),
-			createPeer(addr{ipv6: true, addr: "2620:fe::3"}),
-			createPeer(addr{ipv6: true, addr: "2a02:6b8::4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Equal(t, filtered, peers)
-	})
-
-	t.Run("Overrepresented IPv6 ASN", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: true, addr: "2001:4860:4860::1"}),
-			createPeer(addr{ipv6: true, addr: "2001:4860:4860::2"}),
-			createPeer(addr{ipv6: true, addr: "2001:4860:4860::3"}),
-			createPeer(addr{ipv6: true, addr: "2a02:6b8::4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Len(t, filtered, 1)
-		require.Equal(t, peers[3], filtered[0])
-	})
-
-	t.Run("IPv4+IPv6 acceptable representation", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: false, addr: "1.1.1.1"}, addr{ipv6: true, addr: "2001:4860:4860::1"}),
-			createPeer(addr{ipv6: false, addr: "2.2.2.2"}, addr{ipv6: true, addr: "2001:4860:4860::2"}),
-			createPeer(addr{ipv6: false, addr: "2.2.3.3"}, addr{ipv6: true, addr: "2620:fe::3"}),
-			createPeer(addr{ipv6: false, addr: "4.4.4.4"}, addr{ipv6: true, addr: "2a02:6b8::4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Equal(t, filtered, peers)
-	})
-
-	t.Run("IPv4+IPv6 overrepresentation", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: false, addr: "1.1.1.1"}, addr{ipv6: true, addr: "2001:4860:4860::1"}),
-			createPeer(addr{ipv6: false, addr: "1.1.1.2"}, addr{ipv6: true, addr: "2606:4700:4700::2"}),
-			createPeer(addr{ipv6: false, addr: "1.1.3.3"}, addr{ipv6: true, addr: "2620:fe::3"}),
-			createPeer(addr{ipv6: false, addr: "4.4.4.4"}, addr{ipv6: true, addr: "2620:fe::4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Len(t, filtered, 1)
-		require.Equal(t, peers[3], filtered[0])
-	})
-
-	q.maxIPsPerGroup = 0
-	t.Run("Disabled IP diversity filter", func(t *testing.T) {
-		peers := []*peer.AddrInfo{
-			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
-			createPeer(addr{ipv6: false, addr: "1.1.1.2"}),
-			createPeer(addr{ipv6: false, addr: "1.1.3.3"}),
-			createPeer(addr{ipv6: false, addr: "1.1.4.4"}),
-		}
-		filtered := q.filterPeersByIPDiversity(peers)
-		require.Equal(t, filtered, peers)
-	})
 }
