+++ rt_diversity_filter_test.go
@@ -7,8 +7,10 @@ import (
 
 	kb "github.com/libp2p/go-libp2p-kbucket"
 	"github.com/libp2p/go-libp2p-kbucket/peerdiversity"
+	"github.com/libp2p/go-libp2p/core/peer"
 	bhost "github.com/libp2p/go-libp2p/p2p/host/basic"
 	swarmt "github.com/libp2p/go-libp2p/p2p/net/swarm/testing"
+	ma "github.com/multiformats/go-multiaddr"
 
 	"github.com/stretchr/testify/require"
 )
@@ -152,4 +154,130 @@ func TestRoutingTableEndToEndMaxPerTable(t *testing.T) {
 	time.Sleep(1 * time.Second)
 	require.Len(t, d.routingTable.ListPeers(), 3)
 	require.True(t, d.routingTable.Find(d5.self) == "")
+}
+
+func TestFilterPeersByIPDiversity(t *testing.T) {
+	maxIPsPerGroup := 2
+
+	type addr struct {
+		ipv6 bool
+		addr string
+	}
+	createPeer := func(ips ...addr) *peer.AddrInfo {
+		p := &peer.AddrInfo{ID: "", Addrs: make([]ma.Multiaddr, 0, len(ips))}
+		for _, ip := range ips {
+			var a ma.Multiaddr
+			var err error
+			if ip.ipv6 {
+				a, err = ma.NewMultiaddr("/ip6/" + ip.addr + "/tcp/4001")
+			} else {
+				a, err = ma.NewMultiaddr("/ip4/" + ip.addr + "/tcp/4001")
+			}
+			require.NoError(t, err)
+			p.Addrs = append(p.Addrs, a)
+		}
+		return p
+	}
+
+	t.Run("Different IPv4 blocks", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
+			createPeer(addr{ipv6: false, addr: "2.2.2.2"}),
+			createPeer(addr{ipv6: false, addr: "3.3.3.3"}),
+			createPeer(addr{ipv6: false, addr: "4.4.4.4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Equal(t, peers, filtered)
+	})
+
+	t.Run("Same IPv4 block, but acceptable diversity", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
+			createPeer(addr{ipv6: false, addr: "1.1.2.2"}),
+			createPeer(addr{ipv6: false, addr: "3.3.3.3"}),
+			createPeer(addr{ipv6: false, addr: "4.4.4.4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Equal(t, peers, filtered)
+	})
+
+	t.Run("Overrepresented IPv4 block", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
+			createPeer(addr{ipv6: false, addr: "1.1.1.2"}),
+			createPeer(addr{ipv6: false, addr: "1.1.3.3"}),
+			createPeer(addr{ipv6: false, addr: "4.4.4.4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Len(t, filtered, 1)
+		require.Equal(t, peers[3], filtered[0])
+	})
+
+	t.Run("Different IPv6 ASNs", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: true, addr: "2001:4860:4860::1"}),
+			createPeer(addr{ipv6: true, addr: "2606:4700:4700::2"}),
+			createPeer(addr{ipv6: true, addr: "2620:fe::3"}),
+			createPeer(addr{ipv6: true, addr: "2a02:6b8::4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Equal(t, filtered, peers)
+	})
+
+	t.Run("Same IPv6 ASNs, but acceptable diversity", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: true, addr: "2001:4860:4860::1"}),
+			createPeer(addr{ipv6: true, addr: "2001:4860:4860::2"}),
+			createPeer(addr{ipv6: true, addr: "2620:fe::3"}),
+			createPeer(addr{ipv6: true, addr: "2a02:6b8::4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Equal(t, filtered, peers)
+	})
+
+	t.Run("Overrepresented IPv6 ASN", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: true, addr: "2001:4860:4860::1"}),
+			createPeer(addr{ipv6: true, addr: "2001:4860:4860::2"}),
+			createPeer(addr{ipv6: true, addr: "2001:4860:4860::3"}),
+			createPeer(addr{ipv6: true, addr: "2a02:6b8::4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Len(t, filtered, 1)
+		require.Equal(t, peers[3], filtered[0])
+	})
+
+	t.Run("IPv4+IPv6 acceptable representation", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: false, addr: "1.1.1.1"}, addr{ipv6: true, addr: "2001:4860:4860::1"}),
+			createPeer(addr{ipv6: false, addr: "2.2.2.2"}, addr{ipv6: true, addr: "2001:4860:4860::2"}),
+			createPeer(addr{ipv6: false, addr: "2.2.3.3"}, addr{ipv6: true, addr: "2620:fe::3"}),
+			createPeer(addr{ipv6: false, addr: "4.4.4.4"}, addr{ipv6: true, addr: "2a02:6b8::4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Equal(t, filtered, peers)
+	})
+
+	t.Run("IPv4+IPv6 overrepresentation", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: false, addr: "1.1.1.1"}, addr{ipv6: true, addr: "2001:4860:4860::1"}),
+			createPeer(addr{ipv6: false, addr: "1.1.1.2"}, addr{ipv6: true, addr: "2606:4700:4700::2"}),
+			createPeer(addr{ipv6: false, addr: "1.1.3.3"}, addr{ipv6: true, addr: "2620:fe::3"}),
+			createPeer(addr{ipv6: false, addr: "4.4.4.4"}, addr{ipv6: true, addr: "2620:fe::4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, maxIPsPerGroup)
+		require.Len(t, filtered, 1)
+		require.Equal(t, peers[3], filtered[0])
+	})
+
+	t.Run("Disabled IP diversity filter", func(t *testing.T) {
+		peers := []*peer.AddrInfo{
+			createPeer(addr{ipv6: false, addr: "1.1.1.1"}),
+			createPeer(addr{ipv6: false, addr: "1.1.1.2"}),
+			createPeer(addr{ipv6: false, addr: "1.1.3.3"}),
+			createPeer(addr{ipv6: false, addr: "1.1.4.4"}),
+		}
+		filtered := filterPeersByIPDiversity(peers, 0)
+		require.Equal(t, filtered, peers)
+	})
 }
